move web search cost tracking
